Increases timeout to try to mitigate seeming race condition with accepter
The accepter resource fails, but when following up manually the ram
share IS accepted in the account. Untaint the resource and apply,
and everything is clean. This patch extends the sleep timeout in an
attempt to give the aws services enough time to settle so the accepter
does not fail.